requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-phonenumber-field==8.0.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
phonenumbers==9.0.2
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
